autopep8 applied to all py files

+++ pandas/tests/test_internals.py
@@ -27,13 +27,14 @@ TEST_COLS = ['a', 'b', 'c', 'd', 'e', 'f', 'g', 'h']
 N = 10
 
 
-def get_float_ex(cols=['a', 'c', 'e'], dtype = np.float_):
-    floats = get_float_mat(N, len(cols), dtype = dtype).T
+def get_float_ex(cols=['a', 'c', 'e'], dtype=np.float_):
+    floats = get_float_mat(N, len(cols), dtype=dtype).T
     return make_block(floats, cols, TEST_COLS)
 
 
 def get_complex_ex(cols=['h']):
-    complexes = (get_float_mat(N, 1, dtype = np.float_).T * 1j).astype(np.complex128)
+    complexes = (
+        get_float_mat(N, 1, dtype=np.float_).T * 1j).astype(np.complex128)
     return make_block(complexes, cols, TEST_COLS)
 
 
@@ -49,7 +50,7 @@ def get_bool_ex(cols=['f']):
     return make_block(mat.T, cols, TEST_COLS)
 
 
-def get_int_ex(cols=['g'], dtype = np.int_):
+def get_int_ex(cols=['g'], dtype=np.int_):
     mat = randn(N, 1).astype(dtype)
     return make_block(mat.T, cols, TEST_COLS)
 
@@ -57,6 +58,7 @@ def get_int_ex(cols=['g'], dtype = np.int_):
 def get_dt_ex(cols=['h']):
     mat = randn(N, 1).astype(int).astype('M8[ns]')
     return make_block(mat.T, cols, TEST_COLS)
+
 
 def create_blockmanager(blocks):
     l = []
@@ -68,6 +70,7 @@ def create_blockmanager(blocks):
 
     index_sz = blocks[0].values.shape[1]
     return BlockManager(blocks, [items, np.arange(index_sz)])
+
 
 class TestBlock(unittest.TestCase):
 
@@ -81,7 +84,7 @@ class TestBlock(unittest.TestCase):
         self.int_block = get_int_ex()
 
     def test_constructor(self):
-        int32block = get_int_ex(['a'],dtype = np.int32)
+        int32block = get_int_ex(['a'], dtype=np.int32)
         self.assert_(int32block.dtype == np.int32)
 
     def test_pickle(self):
@@ -331,10 +334,11 @@ class TestBlockManager(unittest.TestCase):
 
     def test_as_matrix_float(self):
 
-        mgr = create_blockmanager([get_float_ex(['c'],np.float32), get_float_ex(['d'],np.float16), get_float_ex(['e'],np.float64)])
+        mgr = create_blockmanager([get_float_ex(['c'], np.float32), get_float_ex(['d'], np.float16), get_float_ex(['e'], np.float64)])
         self.assert_(mgr.as_matrix().dtype == np.float64)
 
-        mgr = create_blockmanager([get_float_ex(['c'],np.float32), get_float_ex(['d'],np.float16)])
+        mgr = create_blockmanager([get_float_ex(
+            ['c'], np.float32), get_float_ex(['d'], np.float16)])
         self.assert_(mgr.as_matrix().dtype == np.float32)
 
     def test_as_matrix_int_bool(self):
@@ -342,10 +346,11 @@ class TestBlockManager(unittest.TestCase):
         mgr = create_blockmanager([get_bool_ex(['a']), get_bool_ex(['b'])])
         self.assert_(mgr.as_matrix().dtype == np.bool_)
 
-        mgr = create_blockmanager([get_int_ex(['a'],np.int64), get_int_ex(['b'],np.int64), get_int_ex(['c'],np.int32), get_int_ex(['d'],np.int16), get_int_ex(['e'],np.uint8) ])
+        mgr = create_blockmanager([get_int_ex(['a'], np.int64), get_int_ex(['b'], np.int64), get_int_ex(['c'], np.int32), get_int_ex(['d'], np.int16), get_int_ex(['e'], np.uint8)])
         self.assert_(mgr.as_matrix().dtype == np.int64)
 
-        mgr = create_blockmanager([get_int_ex(['c'],np.int32), get_int_ex(['d'],np.int16), get_int_ex(['e'],np.uint8) ])
+        mgr = create_blockmanager([get_int_ex(['c'], np.int32), get_int_ex(
+            ['d'], np.int16), get_int_ex(['e'], np.uint8)])
         self.assert_(mgr.as_matrix().dtype == np.int32)
 
     def test_as_matrix_datetime(self):
@@ -355,20 +360,20 @@ class TestBlockManager(unittest.TestCase):
     def test_astype(self):
 
         # coerce all
-        mgr = create_blockmanager([get_float_ex(['c'],np.float32), get_float_ex(['d'],np.float16), get_float_ex(['e'],np.float64)])
+        mgr = create_blockmanager([get_float_ex(['c'], np.float32), get_float_ex(['d'], np.float16), get_float_ex(['e'], np.float64)])
 
-        for t in ['float16','float32','float64','int32','int64']:
+        for t in ['float16', 'float32', 'float64', 'int32', 'int64']:
             tmgr = mgr.astype(t)
             self.assert_(tmgr.as_matrix().dtype == np.dtype(t))
 
         # mixed
-        mgr = create_blockmanager([get_obj_ex(['a','b']),get_bool_ex(['c']),get_dt_ex(['d']),get_float_ex(['e'],np.float32), get_float_ex(['f'],np.float16), get_float_ex(['g'],np.float64)])
-        for t in ['float16','float32','float64','int32','int64']:
-            tmgr = mgr.astype(t, raise_on_error = False).get_numeric_data()
+        mgr = create_blockmanager([get_obj_ex(['a', 'b']), get_bool_ex(['c']), get_dt_ex(['d']), get_float_ex(['e'], np.float32), get_float_ex(['f'], np.float16), get_float_ex(['g'], np.float64)])
+        for t in ['float16', 'float32', 'float64', 'int32', 'int64']:
+            tmgr = mgr.astype(t, raise_on_error=False).get_numeric_data()
             self.assert_(tmgr.as_matrix().dtype == np.dtype(t))
 
     def test_convert(self):
-        
+
         def _compare(old_mgr, new_mgr):
             """ compare the blocks, numeric compare ==, object don't """
             old_blocks = set(old_mgr.blocks)
@@ -395,17 +400,18 @@ class TestBlockManager(unittest.TestCase):
         # noops
         mgr = create_blockmanager([get_int_ex(['f']), get_float_ex(['g'])])
         new_mgr = mgr.convert()
-        _compare(mgr,new_mgr)
+        _compare(mgr, new_mgr)
 
-        mgr = create_blockmanager([get_obj_ex(['a','b']), get_int_ex(['f']), get_float_ex(['g'])])
+        mgr = create_blockmanager(
+            [get_obj_ex(['a', 'b']), get_int_ex(['f']), get_float_ex(['g'])])
         new_mgr = mgr.convert()
-        _compare(mgr,new_mgr)
+        _compare(mgr, new_mgr)
 
         # there could atcually be multiple dtypes resulting
-        def _check(new_mgr,block_type, citems):
+        def _check(new_mgr, block_type, citems):
             items = set()
             for b in new_mgr.blocks:
-                if isinstance(b,block_type):
+                if isinstance(b, block_type):
                     for i in list(b.items):
                         items.add(i)
             self.assert_(items == set(citems))
@@ -415,23 +421,25 @@ class TestBlockManager(unittest.TestCase):
         mat[:, 0] = '1'
         mat[:, 1] = '2.'
         mat[:, 2] = 'foo'
-        b = make_block(mat.T, ['a','b','foo'], TEST_COLS)
+        b = make_block(mat.T, ['a', 'b', 'foo'], TEST_COLS)
 
         mgr = create_blockmanager([b, get_int_ex(['f']), get_float_ex(['g'])])
-        new_mgr = mgr.convert(convert_numeric = True)
+        new_mgr = mgr.convert(convert_numeric=True)
 
-        _check(new_mgr,FloatBlock,['b','g'])
-        _check(new_mgr,IntBlock,['a','f'])
+        _check(new_mgr, FloatBlock, ['b', 'g'])
+        _check(new_mgr, IntBlock, ['a', 'f'])
 
-        mgr = create_blockmanager([b, get_int_ex(['f'],np.int32), get_bool_ex(['bool']), get_dt_ex(['dt']), 
-                                   get_int_ex(['i'],np.int64), get_float_ex(['g'],np.float64), get_float_ex(['h'],np.float16)])
-        new_mgr = mgr.convert(convert_numeric = True)
+        mgr = create_blockmanager(
+            [b, get_int_ex(
+                ['f'], np.int32), get_bool_ex(['bool']), get_dt_ex(['dt']),
+             get_int_ex(['i'], np.int64), get_float_ex(['g'], np.float64), get_float_ex(['h'], np.float16)])
+        new_mgr = mgr.convert(convert_numeric=True)
 
-        _check(new_mgr,FloatBlock,['b','g','h'])
-        _check(new_mgr,IntBlock,['a','f','i'])
-        _check(new_mgr,ObjectBlock,['foo'])
-        _check(new_mgr,BoolBlock,['bool'])
-        _check(new_mgr,DatetimeBlock,['dt'])
+        _check(new_mgr, FloatBlock, ['b', 'g', 'h'])
+        _check(new_mgr, IntBlock, ['a', 'f', 'i'])
+        _check(new_mgr, ObjectBlock, ['foo'])
+        _check(new_mgr, BoolBlock, ['bool'])
+        _check(new_mgr, DatetimeBlock, ['dt'])
 
     def test_xs(self):
         pass
